added an error when hitting delete button without selecting a champion

+++ delete-champion.php
@@ -7,18 +7,6 @@ require_once('header.php');
 
 require_once('connection.php');
 
-// if ($_SERVER['REQUEST_METHOD'] != 'POST') {
-
-//     echo "<form method='post' action='delete-champion.php'>";
-//     echo "<table style='border: solid 1px black;'>";
-//     echo "<tbody>";
-//     echo "<tr><td>Champion Name</td><td><input name='champion_name' type='text' size='25'></td></tr>";
-    
-//     echo "<tr><td></td><td><input type='submit' value='Submit'></td></tr>";
-    
-//     echo "</tbody>";
-//     echo "</table>";
-//     echo "</form>";
 
 if (!isset($_GET['champion_name']) && $_SERVER['REQUEST_METHOD'] != 'POST') {
     
@@ -34,8 +22,6 @@ if (!isset($_GET['champion_name']) && $_SERVER['REQUEST_METHOD'] != 'POST') {
     while ($row = $result->fetch_assoc()) {
         echo "<option value='$row[champion_name]'>$row[champion_name]</option>";
     }
-
-    // echo "<tr><td></td><td><input type='submit' value='Submit'></td></tr>";
     
     echo "</select>";
     echo "<tr><td></td><td><input type='submit' value='Delete'></td></tr>";
@@ -45,31 +31,22 @@ if (!isset($_GET['champion_name']) && $_SERVER['REQUEST_METHOD'] != 'POST') {
 
 else {
     
-    // try {
+
+    if (!array_key_exists('champion_name', $_POST)) {
+        die('error!!!!! You did not select a champion');
+    }
+    
     $stmt = $conn->prepare("DELETE FROM champions WHERE champion_name = ?;");
 
     $stmt->bind_param('s', $_POST['champion_name']);
 
         
     $stmt->execute();
-    // } catch (PDOException $e) {
-    //     echo "Error: " . $e->getMessage();
-    //     die();
-    // }
 
     echo "Success";    
 }
 
 ?>
-
-<div>
-    <ul>
-    	<li><a href="add-champion.php">add a champion</a></li>
-    	<li><a href="edit-champion.php">edit a champion</a></li>
-        <li><a href="delete-champion.php">delete a champion</a></li>
-        <li><a href="view-champion.php">view a champion</a></li>
-    </ul>
-</div>
 
 </div>
 </body>
